BackEnd/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
bcrypt==4.3.0
    # via flask-bcrypt
blinker==1.9.0
    # via flask
build==1.3.0
    # via pip-tools
click==8.2.1
    # via
    #   flask
    #   pip-tools
ecdsa==0.19.1
    # via sendgrid
flask==3.1.2
    # via
    #   -r requirements.in
    #   flask-bcrypt
    #   flask-cors
    #   flask-sqlalchemy
flask-bcrypt==1.0.1
    # via -r requirements.in
flask-cors==6.0.1
    # via -r requirements.in
flask-sqlalchemy==3.1.1
    # via -r requirements.in
gunicorn==23.0.0
    # via -r requirements.in
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
markupsafe==3.0.2
    # via
    #   flask
    #   jinja2
    #   werkzeug
packaging==25.0
    # via
    #   build
    #   gunicorn
pip-tools==7.5.0
    # via -r requirements.in
psycopg2-binary==2.9.10
    # via -r requirements.in
pyjwt==2.10.1
    # via -r requirements.in
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
python-dotenv==1.1.1
    # via -r requirements.in
python-http-client==3.3.7
    # via sendgrid
sendgrid==6.12.4
    # via -r requirements.in
six==1.17.0
    # via ecdsa
sqlalchemy==2.0.43
    # via flask-sqlalchemy
typing-extensions==4.15.0
    # via sqlalchemy
werkzeug==3.1.3
    # via
    #   flask
    #   flask-cors
    #   sendgrid
wheel==0.45.1
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
